GHCJS.Electron.Accelerator

Description

FIXME: doc

Documentation

getGlobalShortcut :: IO GlobalShortcut

Get the canonical GlobalShortcut object, i.e.: the value of require(electron).globalShortcut.

acceleratorRegister :: GlobalShortcut -> JSString -> Callback () -> IO ()

Register a callback for the given shortcut string using the given GlobalShortcut object.

newtype KeyQuery

Represents a particular subset of the set of all key combinations.

Constructors

MkKeyQuery

Fields

_keyCombinations :: [KeyCombination]

simpleKQ :: [KeyModifier] -> KeyCode -> KeyQuery

data KeyCombination

A combination of a key and a modifier.

Constructors

MkKeyCombination

Fields

_kcCode :: !KeyCode

_kcModifiers :: [KeyModifier]

data KeyModifier

A key modifier.

Constructors

KeyModCommand

KeyModControl

KeyModAlt

KeyModAltGr

KeyModShift

KeyModSuper
This is the "Windows key" on Linux/Windows keyboards. On Mac keyboards, it is Cmd.

keyModToAccelerator :: KeyModifier -> Text

FIXME: doc

data KeyCode

A key code.

Constructors

KeyCodeLetter !KeyCodeLetter

KeyCodeDigit !KeyCodeDigit

KeyCodeSymbol !KeyCodeSymbol

KeyCodeFunction !KeyCodeFunction

KeyCodeNavigation !KeyCodeNavigation

KeyCodeSpecial !KeyCodeSpecial

KeyCodeMisc !KeyCodeMisc

keyCodeToAccelerator :: KeyCode -> Text

FIXME: doc

data KeyCodeLetter

Key codes that represent letters.

Constructors

KeyA

KeyB

KeyC

KeyD

KeyE

KeyF

KeyG

KeyH

KeyI

KeyJ

KeyK

KeyL

KeyM

KeyN

KeyO

KeyP

KeyQ

KeyR

KeyS

KeyT

KeyU

KeyV

KeyW

KeyX

KeyY

KeyZ

data KeyCodeDigit

Key codes that represent digits.

Constructors

Key0

Key1

Key2

Key3

Key4

Key5

Key6

Key7

Key8

Key9

data KeyCodeSymbol

Key codes that represent symbols.

Constructors

KeyBacktick
The backtick/tilde key.

KeySemicolon
The semicolon/colon key.

KeyQuote
The single/double quote key.

KeyOBracket
The open-square/curly bracket key.

KeyCBracket
The close-square/curly bracket key.

KeyBackslash
The backslash/pipe key.

KeySlash
The slash/question mark key.

KeyPeriod
The period/greater-than key.

KeyComma
The comma/less-than key.

KeyHyphen
The hyphen/underscore key.

KeyEquals
The equals/plus key.

data KeyCodeFunction

Key codes corresponding to the F1-F24 keys ("function" keys).

Constructors

KeyF1

KeyF2

KeyF3

KeyF4

KeyF5

KeyF6

KeyF7

KeyF8

KeyF9

KeyF10

KeyF11

KeyF12

KeyF13

KeyF14

KeyF15

KeyF16

KeyF17

KeyF18

KeyF19

KeyF20

KeyF21

KeyF22

KeyF23

KeyF24

data KeyCodeNavigation

Key codes corresponding to the arrow keys, page up/down, home, and end.

Constructors

KeyUp

KeyDown

KeyLeft

KeyRight

KeyPageUp

KeyPageDown

KeyHome

KeyEnd

data KeyCodeSpecial

Key codes corresponding to special keys, like XF86AudioPlayPause.

Constructors

KeyMediaNext

KeyMediaPrev

KeyMediaStop

KeyMediaPlay

KeyVolumeUp

KeyVolumeDown

KeyVolumeMute

data KeyCodeMisc

Other various key codes.

Constructors

KeySpace

KeyTab

KeyBackspace

KeyDelete

KeyEnter

KeyInsert

KeyEscape

KeyPrintScreen