GHCJS.Electron.App

Description

FIXME: doc

Documentation

getApp :: IO App

Get the current App as returned by require("electron").app.

appListenerOn :: App -> JSString -> Callback (JSString -> Any) -> IO ()

Listen for the given event and run the given callback whenever it occurs.

appListenerOnce :: App -> JSString -> Callback (JSString -> Any) -> IO ()

Listen for the given event and run the given callback exactly once; i.e.: the callback will be run precisely the first time the event occurs after this function is run.

appRemoveAllListeners :: App -> IO ()

Remove all listeners on the given App.

appRemoveAllListenersOnEvent :: App -> JSString -> IO ()

Remove all listeners for the given event on the given App.

appSetMaxListeners :: App -> Int -> IO ()

Set the maximum number of listeners for events on the given App to the given natural number.

appGetMaxListeners :: App -> IO Int

Get the maximum number of listeners for events on the given App, as set by appSetMaxListeners.

appListenerCount :: App -> JSString -> IO Int

Get the number of listeners for the given event on the given App.

appQuit :: App -> IO ()

Try to close all windows. The before-quit event will first be emitted. If all windows are successfully closed, the will-quit event will be emitted and by default the application would be terminated.

This method guarantees all beforeunload and unload handlers are correctly executed. It is possible that a window cancels the quitting by returning false in beforeunload handler.

appTerminate :: App -> IO ()

Quit the application directly; it will not try to close all windows so cleanup code will not run.

appGetAppPath :: App -> IO Path

Returns the current application directory.

appGetPath

Arguments

:: App

-> JSString
Should be one of the following: "home", "appData", "userData", "cache", "userCache", "temp", "userDesktop", "exe", "module"

-> IO Path
The path associated with the given name.

Get a special path with the given name.

appSetPath

Arguments

:: App

-> JSString
Should be one of the following: "home", "appData", "userData", "cache", "userCache", "temp", "userDesktop", "exe", "module"

-> Path
A file path to set; will be created if it does not exist.

-> IO ()

Overrides the path to a special directory or file associated with name. If the path specifies a directory that does not exist, the directory will be created by this method. On failure an Error would throw.

By default web pages' cookies and caches will be stored under userData directory, if you want to change this location, you have to override the userData path before the ready event of app module gets emitted.

appGetVersion :: App -> IO JSString

Returns the version of loaded application. If no version is found in the application's package.json, then the version of current bundle or executable is returned instead.

appGetName :: App -> IO JSString

Returns the name of the loaded application. Usually the name field of package.json is a short lowercase string, according to the NPM module specification, so usually you should also specify the productName field, which is your application's full capitalized name, and it will be preferred over name by Electron.

appResolveProxy

Arguments

:: App

-> URL
The URL to resolve.

-> Callback (Proxy -> IO ())
The callback to call once the URL is resolved.

-> IO ()

Given a URL, calls the given callback with the proxy used for that URL.

appAddRecent

Arguments

:: App

-> Path
A path to add to the recent documents list.

-> IO ()

Adds the given path to the recent documents list. This list is managed by the system; on Windows you can visit the list from the taskbar; on Mac OS you can visit it from the dock menu.

appClearRecent :: App -> IO ()

Clear the recent documents list.

appSetUserTasks

Arguments

:: App

-> JSVal
A list of tasks to which the user tasks will be set.

-> IO ()

Adds tasks to the Tasks category of JumpList on Windows. This API is only available on Windows.

appGetDock :: App -> IO BrowserWindow

FIXME: doc

appSetDock :: App -> BrowserWindow -> IO ()

FIXME: doc FIXME: may not make sense.

appGetCommandLine :: App -> IO CommandLine

FIXME: doc

appSetCommandLine :: App -> CommandLine -> IO ()

FIXME: doc FIXME: may not make sense.

appMakeSingleInstance

Arguments

:: App

-> Callback (JSArray JSString -> JSString -> IO Bool)
FIXME: doc

-> IO Bool

This method makes your application a "Single Instance Application" instead of allowing multiple instances of your app to run. In other words, this will ensure that at most one instance of your application is running at any time, and launching a new instance will simply signal this instance and exit. NOTE: callback type is (args: string[], workingDirectory: string) => bool FIXME: what does the callback do? what does the returned boolean mean?

appSetUserModelId

Arguments

:: App

-> JSString
FIXME: doc

-> IO ()

FIXME: doc