GHCJS.Electron.BrowserWindowProxy

Description

The BrowserWindowProxy object is returned from window.open and provides limited functionality with the child window.

Documentation

browserBlur :: BrowserWindowProxy -> IO ()

Removes focus from the child window.

browserClose :: BrowserWindowProxy -> IO ()

Forcefully closes the child window without calling its unload event.

browserEval

Arguments

:: BrowserWindowProxy

-> JSString
code

-> IO ()

Evaluates the code in the child window.

browserFocus :: BrowserWindowProxy -> IO ()

Focuses the child window (brings the window to front).

browserPrint :: BrowserWindowProxy -> IO ()

Invokes the print dialog on the child window.

browserPostMessage

Arguments

:: BrowserWindowProxy

-> JSString
message

-> JSString
targetOrigin

-> IO ()

Send a message to the child window with the specified origin or use "*" for no origin preference.

browserIsClosed :: BrowserWindowProxy -> IO Bool

Check if the child window has been closed.