GHCJS.Electron.Clipboard

Description

FIXME: doc

Documentation

getClipboard :: IO Clipboard

Get the global clipboard object.

clipboardReadText

Arguments

:: Clipboard
The clipboard object.

-> JSString
The clipboard type, e.g.: "selection".

-> IO JSString
The text in the clipboard.

Get the current clipboard as plain text.

clipboardWriteText

Arguments

:: Clipboard
The clipboard object.

-> JSString
The clipboard type, e.g.: "selection".

-> JSString
The contents to set the clipboard to.

-> IO ()

Set the clipboard to the given plain text string.

clipboardReadHTML

Arguments

:: Clipboard
The clipboard object.

-> JSString
The clipboard type, e.g.: "selection".

-> IO JSString
The HTML in the clipboard.

Get the current clipboard as an HTML string.

clipboardWriteHTML

Arguments

:: Clipboard
The clipboard object.

-> JSString
The clipboard type, e.g.: "selection".

-> JSString
The contents to set the clipboard to.

-> IO ()

Set the clipboard to the given HTML string.

clipboardReadImage

Arguments

:: Clipboard
The clipboard object.

-> JSString
The clipboard type, e.g.: "selection".

-> IO Image
The clipboard image content.

Get the image content in the clipboard as an Image/NativeImage.

clipboardWriteImage

Arguments

:: Clipboard
The clipboard object.

-> JSString
The clipboard type, e.g.: "selection".

-> Image
The contents to set the clipboard to.

-> IO ()

Set the image content in the clipboard to the given Image/NativeImage.

clipboardReadRTF

Arguments

:: Clipboard
The clipboard object.

-> JSString
The clipboard type, e.g.: "selection".

-> IO JSString
The RTF in the clipboard.

Get the current clipboard as RTF.

clipboardWriteRTF

Arguments

:: Clipboard
The clipboard object.

-> JSString
The clipboard type, e.g.: "selection".

-> JSString
The contents to set the clipboard to.

-> IO ()

Set the clipboard to the given RTF string.

clipboardReadBookmark

Arguments

:: Clipboard
The clipboard object.

-> IO Bookmark
The resultant bookmark object.

Returns a JSVal that is a JavaScript object with the following keys: * title, which maps to the name of the bookmark in the clipboard. * url, which maps to the URL of the bookmark in the clipboard. This function only works on Windows and Mac OS.

clipboardWriteBookmark

Arguments

:: Clipboard
The clipboard object.

-> JSString
The clipboard type.

-> JSString
The bookmark title.

-> URL
The bookmark URL.

-> IO ()

Writes a bookmark with the given title and URL into the given clipboard. This function only works on Windows and Mac OS.

clipboardClear

Arguments

:: Clipboard
The clipboard object.

-> JSString
The clipboard type, e.g.: "selection".

-> IO ()

Clear the current clipboard state.

clipboardFormats

Arguments

:: Clipboard
The clipboard object.

-> JSString
The clipboard type, e.g.: "selection".

-> IO JSVal
An array of supported format strings.

Get the supported formats for the given clipboard.