GHCJS.Electron.Dialog

Documentation

getDialog :: IO Dialog

Get the value of require(electron).dialog.

dialogShowOpen

Arguments

:: Dialog

-> BrowserWindow
Window on which to open the dialog.

-> JSVal
Options object.

-> IO (JSArray Path)
An array of file paths from the user.

Show an "open file" dialog to the user.

dialogShowSave

Arguments

:: Dialog

-> BrowserWindow
Window on which to open the dialog.

-> JSVal
Options object.

-> IO Path
The file path chosen for saving.

Show an "save file" dialog to the user.

dialogShowMessageBox

Arguments

:: Dialog

-> BrowserWindow
Window on which to open the box.

-> JSVal
Options object.

-> IO Int
The index of the clicked button.

This function shows a message box. Note that it will block the process until the message box is closed.

dialogShowErrorBox

Arguments

:: Dialog

-> JSString
The title to display in the error box.

-> JSString
The text content of the error box.

-> IO ()

Display a modal dialog box for error message reporting.

If this is called before the GUI is ready, the message may be emitted stderr instead, especially on Linux. To mitigate this, avoid using this function before the ready event when running on Linux.

dialogCertDialog

Arguments

:: Dialog

-> BrowserWindow
Window on which to open the box.

-> JSVal
Options object.

-> Callback (IO ())
A callback to run once done.

-> IO ()

This function displays a modal dialog that shows a message and certificate information, and gives the user the option of trusting/importing the shown certificate.

NOTE: this function only works on Mac OS