GHCJS.Electron.IPCMain

Description

FIXME: doc

Documentation

getIPCMain :: IO IPCMain

Get the global IPCMain object.

ipcMainListenerOn :: IPCMain -> JSString -> Callback (Event -> Any -> ()) -> IO ()

Listen for the given event and run the given callback whenever it occurs.

ipcMainListenerOnce :: IPCMain -> JSString -> Callback (Event -> Any -> ()) -> IO ()

Listen for the given event and run the given callback exactly once; i.e.: the callback will be run precisely the first time the event occurs after this function is run.

ipcMainRemoveAllListeners :: IPCMain -> IO ()

Remove all listeners on the given IPCMain.

ipcMainRemoveAllListenersOnEvent :: IPCMain -> JSString -> IO ()

Remove all listeners for the given event on the given IPCMain.