GHCJS.Electron.Tray

Description

FIXME: doc

Documentation

data Platform

Constructors

PlatLinux

PlatMacOS

PlatWindows

type Linux = PlatLinux

type Win32 = PlatWindows

type Darwin = PlatMacOS

data KV :: Symbol -> k -> Type where

Constructors

ProxyKV :: KV key val

type (▶) k v = KV k v

data Arg :: Symbol -> k -> Type where

Constructors

ProxyArg :: Arg name ty

type (↦) k v = Arg k v

data Array ty

Constructors

ProxyArray

data Optional ty

Constructors

ProxyOptional

data tyA ∪ tyB

Constructors

ProxyUnion

data Event supportedPlatforms arguments

type TrayBalloonOptions = '["icon" ▶ Optional (Image ∪ JSString), "title" ▶ Optional JSString, "content" ▶ Optional JSString]

type TrayClickEvent = '["altKey" ▶ Bool, "shiftKey" ▶ Bool, "ctrlKey" ▶ Bool, "metaKey" ▶ Bool]

type Point = '["x" ▶ Int, "y" ▶ Int]

type Rectangle = '["x" ▶ Int, "y" ▶ Int, "width" ▶ Int, "height" ▶ Int]

type TrayEvent = '["click" ▶ '('[Linux, Darwin, Win32], '["event" ↦ TrayClickEvent, "bounds" ↦ Rectangle]), "right-click" ▶ '('[Darwin, Win32], '["event" ↦ TrayClickEvent, "bounds" ↦ Rectangle]), "double-click" ▶ '('[Darwin, Win32], '["event" ↦ TrayClickEvent, "bounds" ↦ Rectangle]), "balloon-show" ▶ '('[Win32], '[]), "balloon-click" ▶ '('[Win32], '[]), "balloon-closed" ▶ '('[Win32], '[]), "drop" ▶ '('[Darwin], '[]), "drop-files" ▶ '('[Darwin], '["event" ↦ '[], "files" ↦ Array JSString]), "drop-text" ▶ '('[Darwin], '["event" ↦ '[], "text" ↦ JSString]), "drag-enter" ▶ '('[Darwin], '[]), "drag-leave" ▶ '('[Darwin], '[]), "drag-end" ▶ '('[Darwin], '[])]

trayNew :: Path -> IO Tray

Initialize a new Tray.

trayEventEmitter :: Tray -> IO (EventEmitter TrayEvent)

Cast a Tray to an 'EventEmitter TrayEvent'.

trayDestroy :: Tray -> IO ()

Destroy the given Tray, freeing the associated resources.

traySetImage

Arguments

:: Tray
The tray to modify.

-> Image
The image to set as the tray icon.

-> IO ()

Set the tray icon to the given Image.

traySetImagePath

Arguments

:: Tray
The tray to modify.

-> Path
The path to an image to set as the tray icon.

-> IO ()

Set the tray icon to the image file at the given Path.

traySetPressedImage

Arguments

:: Tray
The tray to modify.

-> Image
The image to set as the "pressed" image.

-> IO ()

Set the "pressed" image for the tray to the given Image. NOTE: this function only works on Mac OS

traySetTitle

Arguments

:: Tray
The tray to modify.

-> JSString
The title displayed next to the tray icon.

-> IO ()

Set the "title" to the given JSString. NOTE: this function only works on Mac OS

traySetHighlightMode

Arguments

:: Tray
The tray to modify.

-> JSString
The mode to set. One of the following: * "selection": Highlight the tray icon when it is clicked and also when its context menu is open. This is the default. * "always": Always highlight the tray icon. * "never": Never highlight the tray icon.

-> IO ()

Sets when the tray’s icon background becomes highlighted (in blue).

NOTE: You can use highlightMode with a BrowserWindow by toggling between "never" and "always" modes when the window visibility changes.

NOTE: this function only works on Mac OS

trayDisplayBalloon

Arguments

:: Tray
The tray to modify.

-> JSString
The title displayed next to the tray icon.

-> IO ()

Create a balloon popup ala Windows XP. NOTE: this function only works on Windows